removed IIIFImagePreview used for testing and replaced with WorkMedia
