+++ common/views/components/IIIFImagePreview/README.md
